+++ hole-punching-tests/src/main.rs
@@ -294,9 +294,7 @@ impl RedisClient {
 
         tracing::debug!("Pushing {key}={value} to redis");
 
-        self.inner.rpush(key, value).await?;
-
-        Ok(())
+        self.inner.rpush(key, value).await.map_err(Into::into)
     }
 
     async fn pop<V>(&mut self, key: &str) -> Result<V>
@@ -308,7 +306,7 @@ impl RedisClient {
 
         let value = self
             .inner
-            .blpop::<_, HashMap<String, String>>(key, 0)
+            .blpop::<_, HashMap<String, String>>(key, 0.0)
             .await?
             .remove(key)
             .with_context(|| format!("Failed to get value for {key} from redis"))?