+++ swarm/src/connection.rs
@@ -31,8 +31,7 @@ pub use supported_protocols::SupportedProtocols;
 
 use crate::handler::{
     AddressChange, ConnectionEvent, ConnectionHandler, DialUpgradeError, FullyNegotiatedInbound,
-    FullyNegotiatedOutbound, ListenUpgradeError, ProtocolSupport, ProtocolsAdded, ProtocolsChange,
-    UpgradeInfoSend,
+    FullyNegotiatedOutbound, ListenUpgradeError, ProtocolSupport, ProtocolsChange, UpgradeInfoSend,
 };
 use crate::stream::ActiveStreamCounter;
 use crate::upgrade::{InboundUpgradeSend, OutboundUpgradeSend};
@@ -51,7 +50,7 @@ use libp2p_core::upgrade;
 use libp2p_core::upgrade::{NegotiationError, ProtocolError};
 use libp2p_core::Endpoint;
 use libp2p_identity::PeerId;
-use std::collections::HashSet;
+use std::collections::{HashMap, HashSet};
 use std::fmt::{Display, Formatter};
 use std::future::Future;
 use std::sync::atomic::{AtomicUsize, Ordering};
@@ -153,8 +152,11 @@ where
         SubstreamRequested<THandler::OutboundOpenInfo, THandler::OutboundProtocol>,
     >,
 
-    local_supported_protocols: HashSet<StreamProtocol>,
+    local_supported_protocols:
+        HashMap<AsStrHashEq<<THandler::InboundProtocol as UpgradeInfoSend>::Info>, bool>,
     remote_supported_protocols: HashSet<StreamProtocol>,
+    protocol_buffer: Vec<StreamProtocol>,
+
     idle_timeout: Duration,
     stream_counter: ActiveStreamCounter,
 }
@@ -187,11 +189,17 @@ where
         idle_timeout: Duration,
     ) -> Self {
         let initial_protocols = gather_supported_protocols(&handler);
+        let mut buffer = Vec::new();
+
         if !initial_protocols.is_empty() {
             handler.on_connection_event(ConnectionEvent::LocalProtocolsChange(
-                ProtocolsChange::Added(ProtocolsAdded::from_set(&initial_protocols)),
+                ProtocolsChange::from_initial_protocols(
+                    initial_protocols.keys().map(|e| &e.0),
+                    &mut buffer,
+                ),
             ));
         }
+
         Connection {
             muxing: muxer,
             handler,
@@ -203,6 +211,7 @@ where
             requested_substreams: Default::default(),
             local_supported_protocols: initial_protocols,
             remote_supported_protocols: Default::default(),
+            protocol_buffer: buffer,
             idle_timeout,
             stream_counter: ActiveStreamCounter::default(),
         }
@@ -250,6 +259,7 @@ where
             substream_upgrade_protocol_override,
             local_supported_protocols: supported_protocols,
             remote_supported_protocols,
+            protocol_buffer,
             idle_timeout,
             stream_counter,
             ..
@@ -287,25 +297,24 @@ where
                     ProtocolSupport::Added(protocols),
                 )) => {
                     if let Some(added) =
-                        ProtocolsChange::add(remote_supported_protocols, &protocols)
+                        ProtocolsChange::add(remote_supported_protocols, protocols, protocol_buffer)
                     {
                         handler.on_connection_event(ConnectionEvent::RemoteProtocolsChange(added));
-                        remote_supported_protocols.extend(protocols);
+                        remote_supported_protocols.extend(protocol_buffer.drain(..));
                     }
-
                     continue;
                 }
                 Poll::Ready(ConnectionHandlerEvent::ReportRemoteProtocols(
                     ProtocolSupport::Removed(protocols),
                 )) => {
-                    if let Some(removed) =
-                        ProtocolsChange::remove(remote_supported_protocols, &protocols)
-                    {
+                    if let Some(removed) = ProtocolsChange::remove(
+                        remote_supported_protocols,
+                        protocols,
+                        protocol_buffer,
+                    ) {
                         handler
                             .on_connection_event(ConnectionEvent::RemoteProtocolsChange(removed));
-                        remote_supported_protocols.retain(|p| !protocols.contains(p));
                     }
-
                     continue;
                 }
             }
@@ -431,16 +440,16 @@ where
                 }
             }
 
-            let new_protocols = gather_supported_protocols(handler);
-            let changes = ProtocolsChange::from_full_sets(supported_protocols, &new_protocols);
+            let changes = ProtocolsChange::from_full_sets(
+                supported_protocols,
+                handler.listen_protocol().upgrade().protocol_info(),
+                protocol_buffer,
+            );
 
             if !changes.is_empty() {
                 for change in changes {
                     handler.on_connection_event(ConnectionEvent::LocalProtocolsChange(change));
                 }
-
-                *supported_protocols = new_protocols;
-
                 continue; // Go back to the top, handler can potentially make progress again.
             }
 
@@ -454,12 +463,14 @@ where
     }
 }
 
-fn gather_supported_protocols(handler: &impl ConnectionHandler) -> HashSet<StreamProtocol> {
+fn gather_supported_protocols<C: ConnectionHandler>(
+    handler: &C,
+) -> HashMap<AsStrHashEq<<C::InboundProtocol as UpgradeInfoSend>::Info>, bool> {
     handler
         .listen_protocol()
         .upgrade()
         .protocol_info()
-        .filter_map(|i| StreamProtocol::try_from_owned(i.as_ref().to_owned()).ok())
+        .map(|info| (AsStrHashEq(info), true))
         .collect()
 }
 
@@ -732,6 +743,25 @@ enum Shutdown {
     Asap,
     /// A shut down is planned for when a `Delay` has elapsed.
     Later(Delay),
+}
+
+// Structure used to avoid allocations when storing the protocols in the `HashMap.
+// Instead of allocating a new `String` for the key,
+// we use `T::as_ref()` in `Hash`, `Eq` and `PartialEq` requirements.
+pub(crate) struct AsStrHashEq<T>(pub(crate) T);
+
+impl<T: AsRef<str>> Eq for AsStrHashEq<T> {}
+
+impl<T: AsRef<str>> PartialEq for AsStrHashEq<T> {
+    fn eq(&self, other: &Self) -> bool {
+        self.0.as_ref() == other.0.as_ref()
+    }
+}
+
+impl<T: AsRef<str>> std::hash::Hash for AsStrHashEq<T> {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        self.0.as_ref().hash(state)
+    }
 }
 
 #[cfg(test)]