+++ protocols/kad/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.46.0
 
+- Included multiaddresses of found peers alongside peer IDs in `GetClosestPeers` query results.
+  See [PR 5475](https://github.com/libp2p/rust-libp2p/pull/5475)
 - Changed `FIND_NODE` response: now includes a list of closest peers when querying the recipient peer ID. Previously, this request yielded an empty response.
   See [PR 5270](https://github.com/libp2p/rust-libp2p/pull/5270)
 - Update to DHT republish interval and expiration time defaults to 22h and 48h respectively, rationale in [libp2p/specs#451](https://github.com/libp2p/specs/pull/451)
@@ -19,6 +21,12 @@
   See [PR 5317](https://github.com/libp2p/rust-libp2p/pull/5317).
 - Use `web-time` instead of `instant`.
   See [PR 5347](https://github.com/libp2p/rust-libp2p/pull/5347).
+- Correctly handle the `NoKnownPeers` error on automatic bootstrap.
+  See [PR 5349](https://github.com/libp2p/rust-libp2p/pull/5349).
+- Improve automatic bootstrap triggering conditions:
+  trigger when the routing table is updated and we have less that `K_VALUE` peers in it,
+  trigger when a new listen address is discovered and we have no connected peers.
+  See [PR 5474](https://github.com/libp2p/rust-libp2p/pull/5474).
 
 ## 0.45.3